feat(nsc): move metrics logic to separate file
